Add tempest to disabled services
Devstack does not respect the `INSTALL_TEMPEST=False` setting and still
tries to install tempest when it is found in the ENABLED_SERVICES
variable. We need to remove it from the list in order to ensure it won't
be installed.

See https://github.com/gophercloud/gophercloud/issues/2408 for more
info.