Don't require to shell-escape conf_overrides
Use heredoc syntax to append the value of `conf_overrides` to the
`local.conf` file. Compared to the previous solution using `echo`, this
removes the need to shell escape your value.

Also remove the useless test, since the default value for
`inputs.conf_overrides` is an empty string which results in a noop.

Fixes #19.